created a separate file for state machine manipulations
