+++ defold-rive/src/res_rive_data.h
@@ -21,6 +21,7 @@
 namespace rive
 {
     class File;
+    class ArtboardInstance;
 }
 
 namespace dmRive
